Editar y eliminar SweetAlert

+++ burger-queen/src/app/admin/workers/workers.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Worker, CreateWorker } from 'src/app/interfaces/workers-interface';
 import { UsersServiceService } from 'src/app/services/users-service.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-workers',
@@ -47,6 +48,7 @@ export class WorkersComponent implements OnInit {
   onAddWorker(newWorker: CreateWorker): void {
     if (!newWorker.name || !newWorker.email || !newWorker.role) {
       console.log('Por favor, completa todos los campos del formulario.');
+      Swal.fire('Error', 'Debes completar todos los campos.');
       return;
   }
   this.workersService.addWorker(newWorker).subscribe(
@@ -81,15 +83,33 @@ onEditWorker(updatedWorker: Worker): void {
 }
 
 onDeleteWorker(worker: Worker): void {
-  this.workersService.deleteWorker(worker.id).subscribe(
-    (res) => {
-      this.workers = this.workers.filter((w) => w.id !== worker.id);
-      console.log("Se elimino", res, worker);
-    },
-    (error) => {
-      console.log("error al eliminar", error)
+  Swal.fire({
+    title: '¿Seguro que deseas eliminar los datos de este empleado?',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Si',
+    cancelButtonText: 'No',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.workersService.deleteWorker(worker.id).subscribe(
+        (res) => {
+          this.workers = this.workers.filter((w) => w.id !== worker.id);
+          console.log("Se elimino", res, worker);
+          Swal.fire(
+            'Listo!',
+            'Se eliminó correctamente la información del empleado.',
+            'success'
+          )
+        },
+        (error) => {
+          console.log("error al eliminar", error)
+          Swal.fire('Error', 'No se eliminó la información', error);
+        }
+      )
     }
-  )
+  })
 }
 
 goToProducts() {
